+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/collections/array/LongArray.kt
@@ -10,32 +10,32 @@ import csense.kotlin.extensions.collections.generic.*
  * @receiver
  * @param receiver (T) -> U
  */
-inline fun <U> LongArray.forEachDiscard(receiver: Function1<Long, U>) =
-        ForeachDiscardResult(count(), this::get, receiver)
+inline fun <U> LongArray.forEachDiscard(receiver: Function1<Long, U>): Unit =
+        GenericArray.foreachDiscardResult(count(), this::get, receiver)
 
 /**
  * Fills this array with the given value
  */
-fun LongArray.fill(value: Long) =
-        fillArray(count(), value, this::set)
+fun LongArray.fill(value: Long): Unit =
+        GenericArray.fillArray(count(), value, this::set)
 
 
 //region Generic collection extensions
 /**
  * Performs traversal in pairs of 2  (with the first index as well)
  */
-inline fun LongArray.forEach2Indexed(action: Function2IndexedUnit<Long, Long>) =
+inline fun LongArray.forEach2Indexed(action: Function2IndexedUnit<Long, Long>): Unit =
         GenericCollectionExtensions.forEach2Indexed(count(), ::elementAt, action)
 
 /**
  * Performs traversal in pairs of 2
  */
-inline fun LongArray.forEach2(action: Function2Unit<Long, Long>) =
+inline fun LongArray.forEach2(action: Function2Unit<Long, Long>): Unit =
         GenericCollectionExtensions.forEach2(count(), ::elementAt, action)
 
 /**
  * Performs backwards traversal on this list.
  */
-inline fun LongArray.forEachBackwards(action: FunctionUnit<Long>) =
+inline fun LongArray.forEachBackwards(action: FunctionUnit<Long>): Unit =
         GenericCollectionExtensions.forEachBackwards(count(), this::elementAt, action)
 //endregion