+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/collections/array/generic/GenericArray.kt
@@ -1,0 +1,3 @@
+package csense.kotlin.extensions.collections.array.generic
+
+object GenericArray