+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/generic/Generic.kt
@@ -1,0 +1,3 @@
+package csense.kotlin.extensions.generic
+
+object Generic