+++ csense-kotlin-common/src/test/kotlin/csense/kotlin/extensions/collections/array/SharedArrayTests.kt
@@ -11,14 +11,14 @@ class SharedArrayTests {
         val count = 10
         var runningCounter = 0
 
-        fillArray(count, 5) { index, value ->
+        GenericArray.fillArray(count, 5) { index, value ->
             index.assert(runningCounter, "index and the running counter should be equal.")
             runningCounter += 1
             value.assert(5, "should be the initial value")
         }
         runningCounter.assert(10, "should run each count, starting from 0")
 
-        fillArray(0, 0) { _, _ ->
+        GenericArray.fillArray(0, 0) { _, _ ->
             failTest("should not run when count is empty / 0")
         }
     }